:art: Refactor the log submodules #13
