+++ logparser/logs/custom/logs.py
@@ -13,9 +13,22 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-"""LogParser package."""
+"""Create the dictionary for log functions related to custom logs.
 
-__version__ = "1.3a0"
-__license__ = "Apache"
-__copyright__ = "Copyright 2016 Real-Time Innovations, Inc."
-__all__ = ["devices", "logparser", "logs", "utils"]
+Functions:
+  + get_regex_list: Get the regular expressions and function list.
+
+Constants:
+  + CUSTOM_PREFIX: Prefix for custom logs.
+"""
+from __future__ import absolute_import
+import logparser.logs.custom.custom as custom
+
+CUSTOM_PREFIX = "#Custom: "
+
+
+def get_regex_list():
+    """Return the regular expressions and functions list for this module."""
+    regex = []
+    regex.append([custom.on_custom_log, CUSTOM_PREFIX + "(.*)"])
+    return regex
