+++ logparser/logs/logs.py
@@ -21,11 +21,11 @@ Functions:
 """
 from __future__ import absolute_import
 import re
-from logparser.custom.logs import get_regex_list as custom_regex
-from logparser.debug.logs import get_regex_list as debug_regex
-from logparser.events.logs import get_regex_list as events_regex
-from logparser.network.logs import get_regex_list as network_regex
-from logparser.routing.logs import get_regex_list as routing_regex
+from logparser.logs.custom.logs import get_regex_list as custom_regex
+from logparser.logs.debug.logs import get_regex_list as debug_regex
+from logparser.logs.events.logs import get_regex_list as events_regex
+from logparser.logs.network.logs import get_regex_list as network_regex
+from logparser.logs.routing.logs import get_regex_list as routing_regex
 
 
 def add_regex(log_list, method, regex):
